engine/requirements.txt
# ===========================================
# ResumeScore Engine - Python Dependencies
# ===========================================

# FastAPI Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# LangChain Core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.10

# PDF/Document Processing
pypdf>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Web Scraping / URL Fetching
aiohttp>=3.9.0
beautifulsoup4>=4.12.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
httpx==0.26.0

# Optional: For better async support
anyio==4.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
